requirements.txt
feed-utils==0.3.9
requests
selenium
beautifulsoup4
hazelcast-python-client
kafka-python
flask-classy
